backend/requirements.txt
authlib==1.3.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.2.0
alembic>=1.13.0
python-dotenv>=1.0.0
redis>=5.0.0
email-validator>=2.0.0
psycopg2-binary>=2.9.9
stripe>=8.0.0